Pull syncGroupPreviews into own file
