+++ packages/shared/src/store/sync/index.ts
@@ -1,4 +1,5 @@
 export * from './sync';
+export { SyncPriority, syncQueue } from '../syncQueue';
 export { syncContacts } from './syncContacts';
 export { syncGroup } from './syncGroup';
-export { SyncPriority, syncQueue } from '../syncQueue';
+export { syncGroupPreviews } from './syncGroupPreviews';
